+++ pl_bolts/models/regression/logistic_regression.py
@@ -36,7 +36,7 @@ class LogisticRegression(pl.LightningModule):
         self.save_hyperparameters()
         self.optimizer = optimizer
 
-        self.linear = nn.Linear(in_features=self.hparams.input_dim, out_features=self.hparams.num_classes, bias=bias).double()
+        self.linear = nn.Linear(in_features=self.hparams.input_dim, out_features=self.hparams.num_classes, bias=bias)
 
     def forward(self, x):
         y_hat = self.linear(x)
